Better handling of beta/release candidate builds
Identify beta and release candidate builds by a specific -suffix, rather
than looking for certain version-specific strings.

Display the specific version string in the output for these builds.

Don't flag a weird build string as "old version" -- flag as unparseable instead

Note for the user that beta or release candidate builds may have problems,
and that they should upgrade to a releae build as soon as one is available.