Raise severity of 'not admin' if render lag present
If render lag is 3% or greater, make the 'not admin' message a warning.

Also change the thresholds on render lag messages slightly, based on my
fairly subjective observations of what levels people consider to be a
problem, and what levels they consider to be a huge problem. I think the
'warning' threshold might still be too high, but we'll see.

Old thresholds:
  Info: pct < 5%
  Warning: 5% <= pct < 15%
  Critical: pct >= 15%

New thresholds:
  Info: pct < 3%
  Warning: 3% <= pct < 10%
  Critical: pct >= 10%