don't crash out on misformated audio buffering lines
Because we don't have a mutex on log output, lines can sometimes get
mixed together. If this happened for an "adding XX milliseconds of audio
buffering" line, the analyzer would crash out because the millisecond count
wasn't in the string position it should have been.

We now parse with a regex, and give ourselves a little bit of framework for
doing fancier things in the future.