Fix Encoder overload detection on multiple encodes
When starting multiple encoding sessions with NVENC (or other hardware
encoder) and receiving encoder overload, it was still reported as
"Encoder Overload", and not specifically as "GPU Encoder Overload".

Reason being: We count lines for each encoder. We then add those up and
check if the count is greater then 1. This check is intended for
mixed-encoder logs (i.e. x264 + any hardware encoder). However as 2
encoding sessions of any encoder is already greater then 1, we get
caught on the first case and never get to the further cases.