windows: Fix GPU check for custom OBS versions
When analyzing a log for a custom OBS version, the version string
comparison would result in checkGPU thinking that the version in the log
was less than '23.2.1'. This resulted in using the pre-23.2.1 adapter
indexes, which started at 1 instead of 0. This, in turn, resulted in the
analyzer thinking that there was only one GPU in the system. If OBS is
running on the dGPU, Adapter 1 would often be the iGPU. If this iGPU was
an Intel iGPU, the analyzer would emit an "Integrated GPU" warning.

Instead of basing the adapter indices on the OBS version, which may have
required processing the OBS version string or adding new version string
processing functions, we can simply try adapter indexes 0-2 and treat
index 0 as if it were possible that it doesn't exist with a try-except
on each call as we already did for one call in each branch. This
resolves the erroneous "Integrated GPU" warning when processing a log
from a "Custom OBS Build" or any other version that results in its
version string evaluating as less than 23.2.1.