windows: Fix check32bitOn64bit
For OBS Studio 29.1.0-beta1, 64-bit OBS did not output "64-bit" in its
logs. While this was fixed for OBS Studio 29.1.0-beta2, the check as it
is currently would probably also catch a hypothetical native WoA
(Windows-on-ARM) build. Additionally, the change in this commit is what
the live log analyzer has been running for about a year. Checking for
the presence of the undesired indicator ("32-bit" or "32bit") seems more
correct than checking for the absence of the desired indicator ("64-bit"
or "64bit").