linux: Detect unofficial Flatpaks
This is a quick and dirty detection that assumes that if a distribution
were to package their own Flatpak, they would also use their own custom
built runtime (we already have one such example).

A "proper" detection would involve changing how and what OBS logs about
the Flatpak environment and use that instead. Since that requires some
additional considerations, that is out of scope for now.